little fixes + new note button open new tab
