+++ src/components/pages/personalSpace.js
@@ -1,20 +1,18 @@
-import React, { Component } from 'react'
+import React from 'react';
 import NotesList from "../Layout/NotesList"
 import Sidebar from "../Layout/Sidebar";
 import "./personalSpace.css";
 
-export default class PersonalSpace extends Component {
-
-    render() {
-        let pwButtons = ['Upload', 'New Note']
-        return (
-            <div className="app__page">
-                <Sidebar buttons={pwButtons}></Sidebar>
-                <div className="personalSpace">
-                    <NotesList></NotesList>
-                </div>
+function PersonalSpace() {
+    let pwButtons = ['Upload', 'New Note']
+    return (
+        <div className="app__page">
+            <Sidebar buttons={pwButtons}></Sidebar>
+            <div className="personalSpace">
+                <NotesList></NotesList>
             </div>
-        )
-
-    }
+        </div>
+    )
 }
+export default PersonalSpace
+
